python/requirements.txt
PyInstaller==3.3.1
pypiwin32==223
pywin32==223
robotframework==3.0.4
robotframework-seleniumlibrary==3.1.1